Fix broken acl_access dumper

+++ src/auth/AclMaxUserIp.cc
@@ -153,7 +153,7 @@ ACLMaxUserIP::dump() const
     if (!maximum)
         return sl;
     SBuf s;
-    s.Printf("%lu", (unsigned long int) maximum);
+    s.Printf("%d", maximum);
     sl.push_back(s);
     return sl;
 }
